# [graphics] optimize ocean-near renderer