--- a/game/graphics/opengl_renderer/Shader.cpp
+++ b/game/graphics/opengl_renderer/Shader.cpp
@@ -80,4 +80,5 @@ ShaderLibrary::ShaderLibrary() {
   at(ShaderId::GENERIC) = {"generic"};
   at(ShaderId::OCEAN_TEXTURE) = {"ocean_texture"};
   at(ShaderId::OCEAN_TEXTURE_MIPMAP) = {"ocean_texture_mipmap"};
+  at(ShaderId::OCEAN_COMMON) = {"ocean_common"};
 }