--- a/game/graphics/opengl_renderer/ocean/OceanNear.cpp
+++ b/game/graphics/opengl_renderer/ocean/OceanNear.cpp
@@ -1,4 +1,5 @@
 #include "OceanNear.h"
+#include "third-party/imgui/imgui.h"
 
 OceanNear::OceanNear(const std::string& name, BucketId my_id)
     : BucketRenderer(name, my_id), m_texture_renderer(false), m_direct(name, my_id, 0x4000) {
@@ -8,6 +9,7 @@ OceanNear::OceanNear(const std::string& name, BucketId my_id)
 }
 
 void OceanNear::draw_debug_window() {
+  ImGui::Checkbox("New", &m_use_new);
   m_direct.draw_debug_window();
 }
 
@@ -62,13 +64,20 @@ void OceanNear::render(DmaFollower& dma,
   }
 
   // direct setup
-  m_direct.reset_state();
   {
+    if (m_use_new) {
+      m_common_ocean_renderer.init_for_near();
+    } else {
+      m_direct.reset_state();
+    }
     auto setup = dma.read_and_advance();
     ASSERT(setup.vifcode0().kind == VifCode::Kind::NOP);
     ASSERT(setup.vifcode1().kind == VifCode::Kind::DIRECT);
     ASSERT(setup.size_bytes == 32);
-    m_direct.render_gif(setup.data, 32, render_state, prof);
+    if (m_use_new) {
+    } else {
+      m_direct.render_gif(setup.data, 32, render_state, prof);
+    }
   }
 
   // oofset and base
@@ -113,5 +122,18 @@ void OceanNear::render(DmaFollower& dma,
   while (dma.current_tag_offset() != render_state->next_bucket) {
     dma.read_and_advance();
   }
-  m_direct.flush_pending(render_state, prof);
+
+  if (m_use_new) {
+    m_common_ocean_renderer.flush(render_state, prof);
+  } else {
+    m_direct.flush_pending(render_state, prof);
+  }
+}
+
+void OceanNear::xgkick(u16 addr, SharedRenderState* render_state, ScopedProfilerNode& prof) {
+  if (m_use_new) {
+    m_common_ocean_renderer.kick_from_near((const u8*)&m_vu_data[addr]);
+  } else {
+    m_direct.render_gif((const u8*)&m_vu_data[addr], UINT32_MAX, render_state, prof);
+  }
 }