… game/graphics/opengl_renderer/ocean/OceanTexture_PS2.cpp