Total ocean time was reduced from 7ms to 2.7 ms in the bad spot in misty. Ocean-near now takes around 0.5ms (used to be over 4ms) and has 4 draws.

Ocean mid/far is still not optimized and takes over 2 ms. I plan to use this `CommonOceanRenderer` to process Ocean Mid draws as well.